pyproject.toml
[tool.poetry]
name = "fitness-app"
version = "0.1.0"
description = ""
authors = ["Sean Deery <sdeery14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai-agents = "^0.3.0"
python-dotenv = "^1.1.1"
SQLAlchemy = "^2.0.32"
aiosqlite = "^0.20.0"
asyncpg = "^0.29.0"
mlflow = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
